WEB_BE/requirements.txt
Flask==2.2.2
pyserial